Reorganization + Fix the controller
The controller broke because a package was re-imported (and I don't know why) that contained a character controller with the same class name. This should now be fixed. The project has also been reorganized, most imported assets are now in the imported folder, not just the root asset folder, and there are now empty parents to keep the hierarchy organized and readable. The door looks hideous, but should work when implemented with an animator component.