tweak door anim
Based off of how the animations work, character controller works, and vertical doors specifically, it was really easy to get stuck under a door and be softlocked. Earlier I changed the scripting to have a timer of "don't close while player is still in," but a dedicated player could still get in hypothetically. I changed the animation itself so that any daring speedrunners would have to do frame perfect movement to get softlocked in the collider.